Decisions Changes
Exiled the coring decisions category to the bottom, same with influence.
Added an R56RP Rules Category for rule related shite.

Cleaned up the Coring mechanic decisions:
Added a "gui" option to only decisions that meet the compliance and resistance requirement on the map. This option is enabled by default.
Changed the toggle button into two on / off buttons to avoid confusion.